+++ CHANGELOG.md
@@ -4,7 +4,7 @@ BikeControl Pro is now available - subscribe to the Pro version to support the p
 - Cross Platform license: use the same subscription on all your devices (Windows, macOS, Android and iOS) without additional charge
 - Synchronize and backup your keymap settings across devices
 - Android: Open Assistant
-- Windows & macOS: start any command/Shortcut, e.g. capture a screenshot, screen video, open a specific app, ... the possibilities are endless!
+- Windows, iOS, macOS: start any command/Shortcut, e.g. capture a screenshot, screen video, open a specific app, ... the possibilities are endless!
 - and much more to come in the future!
 
 **Features**: